+++ src/lib_gui/qt/element/code/QtCodeArea.h
@@ -89,7 +89,7 @@ public:
 
 	bool setFocus(Id locationId);
 	bool moveFocus(CodeFocusHandler::Direction direction, size_t lineNumber, Id locationId);
-	bool moveFocusToLine(size_t lineNumber, int targetColumn, bool up);
+	bool moveFocusToLine(int lineNumber, int targetColumn, bool up);
 	bool moveFocusInLine(size_t lineNumber, Id locationId, bool forward);
 	void activateLocationId(Id locationId, bool fromMouse);
 
@@ -107,7 +107,7 @@ protected:
 
 private slots:
 	void updateLineNumberAreaWidth(int newBlockCount = 0);
-	void updateLineNumberArea(QRect , int);
+	void updateLineNumberArea(QRect, int);
 	void setIDECursorPosition();
 	void setCopyAvailable(bool yes);
