+++ src/lib_gui/qt/element/code/QtCodeFileSingle.cpp
@@ -247,7 +247,7 @@ void QtCodeFileSingle::setFocusOnTop()
 {
 	if (m_area)
 	{
-		m_area->moveFocusToLine(m_area->getStartLineNumber(), 0, false);
+		m_area->moveFocusToLine(static_cast<int>(m_area->getStartLineNumber()), 0, false);
 	}
 }
